Adapted the accelerometer API to Cocoa 2.0. In the process also implemented support for Java-interfaces (which are now mapped to Objective-C protocols) and support for primitive type double (which is used as the data type for the Cocoa 2.0 accelerometer readings).
git-svn-id: http://xmlvm.svn.sourceforge.net/svnroot/xmlvm/trunk/xmlvm@141 0874c2aa-48f6-4a0f-8a6b-e32f4465a257

+++ src/test/iphone/org/xmlvm/test/iphone/ifireworks/Gravity.java
@@ -5,6 +5,6 @@ package org.xmlvm.test.iphone.ifireworks;
 public class Gravity
 {
 
-    static float xGV = 0.0f;
-    static float yGV = 0.5f;
+    static double xGV = 0.0f;
+    static double yGV = 0.5f;
 }
